requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
daphne==4.2.0
DateTime==5.5
distlib==0.3.8
Django==5.0.6
django-autoslug==1.9.9
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
fastapi==0.115.12
fastcore==1.7.25
fastlite==0.0.13
filelock==3.15.4
Flask==3.0.3
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.1
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.10.1
python-multipart==0.0.19
pytz==2025.1
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlite-minutils==4.0.3
sqlparse==0.5.0
starlette==0.41.3
threadpoolctl==3.6.0
Twisted==25.5.0
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.32.1
virtualenv==20.26.3
VirtualEnvironment==0.1.2
watchfiles==1.0.0
websockets==14.1
Werkzeug==3.0.3
zope.interface==7.2
